media-plugins/alsa-lib: Initial 1.2.7.1-r1 (also in gentoo)
DO NOT USE

USE="pipewire" patches the source to link with pipewire and handle a
running pipewire server appropriately. USE="pulseaudio" depends on
pulseaudo (NOT libpulse as in gentoo)

The idea is you can have both pulseaudio and pipewire daemons on the
system and choose to one or the other at any time.

TODO: avoid the SIGSEGV when pw_main_loop_new(0) returns NULL in some
situations.  Some applications trigger this.

FIX pulesudio in commit message