dev-util/cargo-c-dist-bin: Initial: 0.10.15 - bindist
- provides dev-util/cargo_c from Gentoo
- TODO fix bindist in ebuild

;madhu 250814 the problem is dev-debug/bpftrace moving to rust in the
default gentoo configuration.

dev-util/cargo_c is a dependency for dev-libs/blazesym_c which gentoo
(for which gentoo ships a 30 mb crates tarball, which
cargo-vendorer-filter for v0.1.3 reduces to 515k).  however the llvm
investors have arranged for one of the transitive dependencies, goblin,
of which 2 versions are included depends on "edition 2024", which puts
both the rust-1.79 and rust-1.82 i have out of the game and forces a
rust upgrade.

also bpftrace-0.23.5 doesn't seem to build on 6.15.x with libbpf-1.6.1